requirement.txt
streamlit 
google-generativeai 
PyMuPDF 
Pillow